+++ src/attributes.ts
@@ -355,6 +355,7 @@ export abstract class AttributesToken extends Token {
 		}
 	}
 
+	/** @private */
 	override escape(): void {
 		LSP: {
 			if (this.type !== 'ext-attrs') {