+++ src/table/td.ts
@@ -317,6 +317,7 @@ export abstract class TdToken extends TableBaseToken {
 		}
 	}
 
+	/** @private */
 	override escape(): void {
 		LSP: {
 			super.escape();