+++ src/tag/html.ts
@@ -372,6 +372,7 @@ export abstract class HtmlToken extends TagToken {
 		return result;
 	}
 
+	/** @private */
 	override getRange(): AstRange | undefined {
 		const {selfClosing, name} = this,
 			{html: [, selfClosingTags, voidTags]} = this.getAttribute('config');